Release 1.10:     - Memory optimization in HyBR. 	- The plugin is updated to use Parallel Colt 0.8.2.
git-svn-id: https://svn.code.sf.net/p/iterativedeconv/code@3 1d76879a-1871-4656-b8f0-ef00bcd114e6

+++ ParallelIterativeDeconvolution/src/edu/emory/mathcs/restoretools/iterative/mrnsd/MRNSDDoubleIterativeDeconvolver2D.java
@@ -122,9 +122,9 @@ public class MRNSDDoubleIterativeDeconvolver2D extends AbstractDoubleIterativeDe
             imX.show();
         }
         int k;
-        rowList = new IntArrayList(B.size() / 2);
-        columnList = new IntArrayList(B.size() / 2);
-        valueList = new DoubleArrayList(B.size() / 2);
+        rowList = new IntArrayList((int)B.size() / 2);
+        columnList = new IntArrayList((int)B.size() / 2);
+        valueList = new DoubleArrayList((int)B.size() / 2);
         for (k = 0; k < maxIters; k++) {
             if (rnrm <= stoppingTol) {
                 IJ.log("MRNSD converged after " + k + "iterations.");
