+++ ParallelIterativeDeconvolution/src/edu/emory/mathcs/restoretools/iterative/mrnsd/MRNSDFloatIterativeDeconvolver3D.java
@@ -124,10 +124,10 @@ public class MRNSDFloatIterativeDeconvolver3D extends AbstractFloatIterativeDeco
             imX.show();
         }
         int k;
-        sliceList = new IntArrayList(B.size() / 2);
-        rowList = new IntArrayList(B.size() / 2);
-        columnList = new IntArrayList(B.size() / 2);
-        valueList = new FloatArrayList(B.size() / 2);
+        sliceList = new IntArrayList((int)B.size() / 2);
+        rowList = new IntArrayList((int)B.size() / 2);
+        columnList = new IntArrayList((int)B.size() / 2);
+        valueList = new FloatArrayList((int)B.size() / 2);
         for (k = 0; k < maxIters; k++) {
             if (rnrm <= stoppingTol) {
                 IJ.log("MRNSD converged after " + k + "iterations.");
