+++ ParallelIterativeDeconvolution/src/edu/emory/mathcs/restoretools/iterative/preconditioner/FFTDoublePreconditioner3D.java
@@ -159,7 +159,7 @@ public class FFTDoublePreconditioner3D implements DoublePreconditioner3D {
             B = new DenseDoubleMatrix3D(imSize[0], imSize[1], imSize[2], (double[]) b.elements(), 0, 0, 0, imSize[1] * imSize[2], imSize[2], 1, false);
         }
         B = solve(B, transpose);
-        return new DenseDoubleMatrix1D(B.size(), (double[]) B.elements(), 0, 1, false);
+        return new DenseDoubleMatrix1D((int)B.size(), (double[]) B.elements(), 0, 1, false);
     }
 
     public DoubleMatrix3D solve(AbstractMatrix3D B, boolean transpose) {
@@ -426,8 +426,8 @@ public class FFTDoublePreconditioner3D implements DoublePreconditioner3D {
     }
 
     private double defaultTol2(DoubleMatrix3D E, DoubleMatrix3D B) {
-        DoubleMatrix1D s = new DenseDoubleMatrix1D(E.size());
-        System.arraycopy((double[]) E.elements(), 0, (double[]) s.elements(), 0, s.size());
+        DoubleMatrix1D s = new DenseDoubleMatrix1D((int)E.size());
+        System.arraycopy((double[]) E.elements(), 0, (double[]) s.elements(), 0, (int)s.size());
         final double[] evalues = (double[]) s.elements();
         IntComparator compDec = new IntComparator() {
             public int compare(int a, int b) {
@@ -442,10 +442,10 @@ public class FFTDoublePreconditioner3D implements DoublePreconditioner3D {
         AbstractMatrix3D Bhat = ((DenseDoubleMatrix3D) B).getFft3();
         ((DComplexMatrix3D) Bhat).assign(DComplexFunctions.abs);
         Bhat = ((DComplexMatrix3D) Bhat).getRealPart();
-        DoubleMatrix1D bhat = new DenseDoubleMatrix1D(Bhat.size(), (double[]) ((DoubleMatrix3D) Bhat).elements(), 0, 1, false);
+        DoubleMatrix1D bhat = new DenseDoubleMatrix1D((int)Bhat.size(), (double[]) ((DoubleMatrix3D) Bhat).elements(), 0, 1, false);
         bhat = bhat.viewSelection(indices);
         bhat.assign(DoubleFunctions.div((double) Math.sqrt(B.size())));
-        int n = s.size();
+        int n = (int)s.size();
         double[] rho = new double[n - 1];
         rho[n - 2] = bhat.getQuick(n - 1) * bhat.getQuick(n - 1);
         DoubleMatrix1D G = new DenseDoubleMatrix1D(n - 1);
