+++ ParallelIterativeDeconvolution/src/edu/emory/mathcs/restoretools/iterative/preconditioner/FFTFloatPreconditioner2D.java
@@ -145,7 +145,7 @@ public class FFTFloatPreconditioner2D implements FloatPreconditioner2D {
             B = new DenseFloatMatrix2D(imSize[0], imSize[1], (float[]) b.elements(), 0, 0, imSize[1], 1, false);
         }
         B = solve(B, transpose);
-        return new DenseFloatMatrix1D(B.size(), (float[]) B.elements(), 0, 1, false);
+        return new DenseFloatMatrix1D((int)B.size(), (float[]) B.elements(), 0, 1, false);
     }
 
     public FloatMatrix2D solve(AbstractMatrix2D B, boolean transpose) {
@@ -387,8 +387,8 @@ public class FFTFloatPreconditioner2D implements FloatPreconditioner2D {
     }
 
     private float defaultTol2(FloatMatrix2D E, FloatMatrix2D B) {
-        FloatMatrix1D s = new DenseFloatMatrix1D(E.size());
-        System.arraycopy((float[]) E.elements(), 0, (float[]) s.elements(), 0, s.size());
+        FloatMatrix1D s = new DenseFloatMatrix1D((int)E.size());
+        System.arraycopy((float[]) E.elements(), 0, (float[]) s.elements(), 0, (int)s.size());
         final float[] evalues = (float[]) s.elements();
         IntComparator compDec = new IntComparator() {
             public int compare(int a, int b) {
@@ -403,10 +403,10 @@ public class FFTFloatPreconditioner2D implements FloatPreconditioner2D {
         AbstractMatrix2D Bhat = ((DenseFloatMatrix2D) B).getFft2();
         ((FComplexMatrix2D) Bhat).assign(FComplexFunctions.abs);
         Bhat = ((FComplexMatrix2D) Bhat).getRealPart();
-        FloatMatrix1D bhat = new DenseFloatMatrix1D(Bhat.size(), (float[]) ((FloatMatrix2D) Bhat).elements(), 0, 1, false);
+        FloatMatrix1D bhat = new DenseFloatMatrix1D((int)Bhat.size(), (float[]) ((FloatMatrix2D) Bhat).elements(), 0, 1, false);
         bhat = bhat.viewSelection(indices);
         bhat.assign(FloatFunctions.div((float) Math.sqrt(B.size())));
-        int n = s.size();
+        int n = (int)s.size();
         float[] rho = new float[n - 1];
         rho[n - 2] = bhat.getQuick(n - 1) * bhat.getQuick(n - 1);
         FloatMatrix1D G = new DenseFloatMatrix1D(n - 1);
