+++ ParallelIterativeDeconvolution/src/edu/emory/mathcs/restoretools/iterative/wpl/WPLFloatIterativeDeconvolver3D.java
@@ -964,7 +964,7 @@ public class WPLFloatIterativeDeconvolver3D {
         final float[] elemsB = (float[]) B.elements();
         final float[] elemsAX = (float[]) AX.elements();
         final float[] elemsX = (float[]) X.elements();
-        final int size = B.size();
+        final int size = (int)B.size();
         int np = ConcurrencyUtils.getNumberOfThreads();
         if ((np > 1) && (size >= ConcurrencyUtils.getThreadsBeginN_3D())) {
             Future<?>[] futures = new Future[np];
@@ -1021,7 +1021,7 @@ public class WPLFloatIterativeDeconvolver3D {
         final float[] x = (float[]) X.elements();
         final float SCALE = (float) (10 / Math.log(10));
         final float minVal = (float) Math.exp(minDB / SCALE);
-        int size = X.size();
+        int size = (int)X.size();
         int np = ConcurrencyUtils.getNumberOfThreads();
         if ((np > 1) && (size >= ConcurrencyUtils.getThreadsBeginN_3D())) {
             Future<?>[] futures = new Future[np];
@@ -1054,7 +1054,7 @@ public class WPLFloatIterativeDeconvolver3D {
     private static float unDB(FloatMatrix3D X) {
         final float[] x = (float[]) X.elements();
         final float SCALE = (float) (10 / Math.log(10));
-        final int size = X.size();
+        final int size = (int)X.size();
         float min = Float.MAX_VALUE;
         int np = ConcurrencyUtils.getNumberOfThreads();
         if ((np > 1) && (size >= ConcurrencyUtils.getThreadsBeginN_3D())) {
